Move service into app
