+++ app/brain/exercises_management/tests/rep_exercises_management_tests.py
@@ -82,7 +82,8 @@ class RepExercisesManagementTests(unittest.TestCase):
         self.exercises[3].id = 4
         self.exercises[4].id = 5
 
-    @patch('app.brain.exercises_management.rep_exercises_management.RepExercisesTaxonomyService.get_list_of_all_exercises')
+    @patch(
+        'app.brain.exercises_management.rep_exercises_management.RepExercisesTaxonomyService.get_list_of_all_exercises')
     def test_get_valid_id_exercise_pairs(self, taxonomy_service_mock):
         taxonomy_service_mock.return_value = self.exercises
         expected_results = [('4', 'a_exercise'), ('3', 'b_exercise'), ('1', 'c_exercise'), ('5', 'd_exercise'),
@@ -114,4 +115,3 @@ class RepExercisesManagementTests(unittest.TestCase):
         self.assertEqual(actual_result.reps, expected_result.reps)
         self.assertEqual(actual_result.weight, expected_result.weight)
         self.assertEqual(actual_result.date, expected_result.date)
-
